Refactored current order into shared hook
