+++ src/pages/public/payment/index.tsx
@@ -1,13 +1,6 @@
-import React, { useEffect, useState } from 'react';
-import { useAppDispatch, useAppSelector } from '../../../store/hooks';
-import {
-	navigate,
-	refreshOrder,
-	selectCurrentOrderId,
-	selectOrders,
-	selectOrdersState
-} from '../../../store/public-store';
-import { IGetOrderResponse } from '@synonymdev/blocktank-client';
+import React, { useEffect } from 'react';
+import { useAppDispatch } from '../../../store/hooks';
+import { navigate } from '../../../store/public-store';
 import FormCard from '../../../components/form-card';
 import Spinner from '../../../components/spinner';
 import './index.scss';
@@ -20,42 +13,11 @@ import { ReactComponent as BitcoinIconActive } from '../../../icons/bitcoin-acti
 import { ReactComponent as BitcoinIcon } from '../../../icons/bitcoin.svg';
 import ErrorPage from '../error';
 import IconRing from '../../../components/icon-ring';
+import useOrder from '../../../hooks/useOrder';
 
 function PaymentPage(): JSX.Element {
-	const [isLoading, setIsLoading] = useState(true);
-	const orderId = useAppSelector(selectCurrentOrderId);
-	const [order, setOrder] = useState<IGetOrderResponse | undefined>(undefined);
-	const orders = useAppSelector(selectOrders);
-	const ordersState = useAppSelector(selectOrdersState);
 	const dispatch = useAppDispatch();
-
-	// TODO move to reusable hook
-	useEffect(() => {
-		const newOrder = orders.find((o) => o._id === orderId);
-		if (newOrder) {
-			setOrder(newOrder);
-			setIsLoading(false);
-		}
-	}, [orders]);
-
-	useEffect(() => {
-		dispatch(refreshOrder(orderId))
-			.catch((e) => alert(e))
-			.finally(() => setIsLoading(false));
-	}, []);
-
-	useEffect(() => {
-		const intervalId = setInterval(() => {
-			if (order?.state === 500) {
-				// Once channel is open stop trying
-				return;
-			}
-
-			dispatch(refreshOrder(orderId)).catch((e) => alert(e));
-		}, 5000);
-
-		return () => clearInterval(intervalId);
-	}, []);
+	const { order, orderState } = useOrder();
 
 	// Once we've received the payment show the success view for 3 seconds before navigating to claim frame
 	useEffect(() => {
@@ -65,7 +27,7 @@ function PaymentPage(): JSX.Element {
 	}, [order]);
 
 	if (!order) {
-		if (ordersState === 'loading' || isLoading) {
+		if (orderState === 'loading' || orderState === 'idle') {
 			return (
 				<FormCard>
 					<Spinner style={{ fontSize: 8 }} centered />
@@ -82,8 +44,6 @@ function PaymentPage(): JSX.Element {
 		btc_address,
 		zero_conf_satvbyte,
 		purchase_invoice,
-		amount_received,
-		price,
 		onchain_payments,
 		order_expiry,
 		state,
@@ -114,6 +74,7 @@ function PaymentPage(): JSX.Element {
 			break;
 		}
 		case 400: // Given up
+		case 410: // Expired
 		case 500: // Channel open
 		case 300: // Channel opening
 		case 450: // Channel closed
